+++ core/test/ConfigurationSpec.scala
@@ -124,11 +124,11 @@ class ConfigurationSpec extends FlatSpec with ShouldMatchers with DefaultConvert
     config5[String]("nums.five") should be ("V")
   }
 
-  it can "be dettach from a configuration" in {
+  it can "be detach from a configuration" in {
     val data2 = Map( "one" -> "1", "two" -> "2" )
     val config2 = Configuration( data2 )
     val config3 = config attach ("nums", config2 )
-    val config4 = config3 dettach ("nums")
+    val config4 = config3 detach ("nums")
     config4 should be (config2)
   }
 
@@ -166,6 +166,15 @@ class ConfigurationObjectSpec extends FlatSpec with ShouldMatchers with io.IOHel
     config[Int]("bazz") should be (2)
   }
 
+  it can "support lists directly with key value pairs" in {
+    val config = Configuration(
+      "foo"-> List(10), 
+      "bar"-> ("hello"::"world"::Nil)
+    )
+    config[List[Int]]("foo") should be (List(10)) 
+    config[List[String]]("bar") should be (List("hello","world")) 
+  }
+
 
   it can "be created from a string using a given format" in {
     val s = 
@@ -180,8 +189,7 @@ class ConfigurationObjectSpec extends FlatSpec with ShouldMatchers with io.IOHel
     config.get[String]("baz") should be (Some("hello world"))
   }
 
-  it can "be loaded from a file" in {
-    val filename = "/tmp/configrity_configuration_obj_spec.conf"
+  it can "be loaded from a file" in {    
     val fmt = FlatFormat
     val s = 
       """
@@ -189,7 +197,7 @@ class ConfigurationObjectSpec extends FlatSpec with ShouldMatchers with io.IOHel
       bar = 2
       baz = "hello world"
       """
-    autoFile( filename, s ){ file =>
+    autoFile( s ){ file =>
       val fn = file.getAbsolutePath
       val config = Configuration.load(fn,fmt)
       config.get[Boolean]("foo") should be (Some(true))
@@ -198,21 +206,8 @@ class ConfigurationObjectSpec extends FlatSpec with ShouldMatchers with io.IOHel
       val config2 = Configuration.load(fn)
       config2 should be (config)
     }
-    /*
-    val writer = new java.io.PrintWriter( filename )
-    writer.println(s)
-    writer.close()
-    try {
-      val config = Configuration.load(filename,fmt)
-      config.get[Boolean]("foo") should be (Some(true))
-      config.get[Int]("bar") should be (Some(2))
-      config.get[String]("baz") should be (Some("hello world"))
-      val config2 = Configuration.load(filename)
-      config2 should be (config)
-    } finally {
-      ( new java.io.File(filename) ).delete
-    }*/
   }
+
 
 
 }